cambios imagen editar

+++ controllers/CoursesController.php
@@ -106,37 +106,41 @@ class CoursesController {
                 if ($validate->isOk()) {
                     $course = new Course($_POST);
                     $course->setId($id);
-                    $course->setImage_link($courseOld->getImage_link());
-                    var_dump($course);
-
-                    /*if ($_FILES["image"]["name"] !== "") {
+                    
+                    if ($_FILES["image"]["name"] != "") {
+                        
                         $tmp = $_FILES["image"]["tmp_name"];
                         $name = $_FILES["image"]["name"];
                         $size = $_FILES["image"]["size"];
                         $type = $_FILES["image"]["type"];
-                        $path = new ConfigPath();
+                        $path = new Config();
+                        
                         $destino = $path->path_imagenes . "/" . $name;
                         if ($type === "image/jpeg") {
                             $ok = move_uploaded_file($tmp, $destino);
                             if ($ok) {
+                                
                                 $course->setImage_link($name);
-                                $file = $path->path_imagenes . "/" . $courseOld->getImage();
+                                $file = $path->path_imagenes . "/" . $courseOld->getImage_link();
                                 if (file_exists($file)) {
                                     unlink($file);
                                 }
-                                $view->image = Config::URL_IMG . "/" . $course->getImage_link();
+
+                                $view->image = Config::PATH_IMG . "/" . $course->getImage_link();
                             }
                         } else {
                             $view->errores["imagen"] = "El archivo seleccionado no es una imagen.";
                         }
                     } else {
-                        $view->imagen = Config::URL_IMG . "/" . $courseOld->getImagen();
-                    }*/
+                        $course->setImage_link($courseOld->getImage_link());
+                        $view->imagen = Config::PATH_IMG . $courseOld->getImage_link();
+                    }
 
                     $id = $this->model->editCourse($course);
                     if ($id !== false) {
-                        header("location:" . Config::URL_BASE . "courses/editCourses");
                         echo "Se ha actualizado el curso";
+                        header("location:" . Config::URL_BASE . "courses/editCourses");
+                        
                     }
                 }
             } else {
@@ -156,7 +160,7 @@ class CoursesController {
                     "image_link" => $courseOld->getImage_link(),
                     "status" => $courseOld->getStatus()
                 ];
-                //$view->image = Config::URL_IMG . "/" . $courseOld->getImagen();
+                $view->image = Config::PATH_IMG . $courseOld->getImage_link();
             }
         } catch (CourseValidatorException $e) {
             $errores = $e->getMessagesErrores();
